requirements.txt
attrs>=17.4,<17.5
flake8
